requirements.txt
# RO Market Crawler - Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP Client
httpx>=0.27.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
aiosqlite>=0.20.0

# Scheduler
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
